House cleaning, finalize QInput*
