+++ src/components/QueryInput/QDebugDevSimple.tsx
