+++ src/components/QueryInput/QInputRange/helpers.ts
