+++ src/components/QueryInput/QInputRange/presetOptions.ts
