+++ src/components/QueryInput/QInputRange/roadmap.txt
@@ -1,0 +1,1 @@
+Deprecate QInputPair in favor of QInputRange.
