+++ src/components/QueryInput/QInputText/QInputText.stories.tsx
