+++ src/components/QueryInput/QInputText/QInputText.test.tsx
