+++ src/components/QueryInput/QInputText/QInputTextStory.tsx
