+++ src/components/QueryInput/term.types.ts
