+++ src/components/QueryInput/types.ts
